+++ packages/terra-framework-docs/src/terra-dev-site/test/compact-interactive-list/WidthBreakpointsHorizontalFlow.test.jsx
@@ -61,19 +61,19 @@ const rows = [
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Col_1',
+    displayName: 'Icon',
     width: '60px',
     align: alignTypes.CENTER,
   },
   {
     id: 'Column-1',
-    displayName: 'Col_2',
+    displayName: 'Service name',
     width: '200px',
     flexGrow: true,
   },
   {
     id: 'Column-2',
-    displayName: 'Col_3',
+    displayName: 'Featured',
     width: '60px',
     align: alignTypes.CENTER,
   },
@@ -87,6 +87,7 @@ const WidthBreakpointsHorizontalFlow = () => {
       rows={rows}
       columns={cols}
       numberOfColumns={getNumberOfColumns(activeBreakpoint)}
+      rowHeaderIndex={1}
       flowHorizontally
     />
   );